时隔一年半修复bug

+++ SerialDebug.h
@@ -2,8 +2,8 @@
   ******************************************************************************
   * @file	 SerialDebug.h
   * @author  Hongxi Wong
-  * @version V1.0.0
-  * @date    2019/12/18
+  * @version V1.0.1
+  * @date    2020/2/17
   * @brief   UWP Serial debug assistant lib based on HAL
   ******************************************************************************
   * @attention
@@ -21,9 +21,9 @@ extern uint8_t DebugBuffer[100];
 extern uint8_t Debug_Count;
 extern uint8_t Debug_Period;
 
-void Serial_Debug(UART_HandleTypeDef *huart, uint8_t debug_period, float a, float b, float c, float d, float e, float f);
+void Serial_Debug(UART_HandleTypeDef *huart, uint16_t debug_period, float a, float b, float c, float d, float e, float f);
 void Debug_Buf_Send(UART_HandleTypeDef *huart, uint8_t *data, uint8_t length);
 void float2char(float floatdata, uint8_t *buffer, uint8_t n, uint8_t *position);
-void Debug_Buf_Generate(UART_HandleTypeDef *huart,float a, float b, float c, float d, float e, float f, uint8_t *data);
+void Debug_Buf_Generate(UART_HandleTypeDef *huart, float a, float b, float c, float d, float e, float f, uint8_t *data);
 
 #endif
